Workaround nfn-agent add-port failure
The nfn-agent requires that ovs be up and running.  Add an
initContainer to wait for ovs to be ready.

Note that it is not sufficient to wait at the user level as joining a
node to the cluster will cause immediate creation of both the ovs and
nfn-agent Pods.

The error in the nfn-agent logs will look something like:

  failure in creating Node OVS internal port - ovn4nfv0-cab2f0: exit
  status 1 - \"ovs-vsctl: unix:/var/run/openvswitch/db.sock: database
  connection failed (No such file or directory)

Signed-off-by: Todd Malsbary <todd.malsbary@intel.com>
Change-Id: Ia68d78ce712d4fcf0c0951df93e1532b9297fd58